Revert "6-04 使用Redux-thunk中间件实现ajax数据请求"
This reverts commit deb6efd3a38384d419619537c8f266db483caf2d.
