+++ package.json
@@ -8,8 +8,7 @@
     "react-dom": "^16.11.0",
     "react-scripts": "3.2.0",
     "react-transition-group": "^4.3.0",
-    "redux": "^4.0.4",
-    "redux-thunk": "^2.3.0"
+    "redux": "^4.0.4"
   },
   "scripts": {
     "start": "react-scripts start",
