+++ src/TodoList.js
@@ -33,22 +33,18 @@ class TodoList extends Component {
 
   componentDidMount () {
     // 模仿 ajax (axios)
-    // new Promise((resolve, reject) => {
-    //   const res = {
-    //     data: ['hello', 'javascript', 'world'],
-    //     status: 200
-    //   };
-    //   resolve(res.data);
-    // })
-    //   .then(response => {
-    //     console.log('response', response);
-    //     const action = actionCreators.initListAction(response);
-    //     store.dispatch(action);
-    //   });
-
-    // 异步获取数据 放入action中。
-    const action = actionCreators.getTodoList();
-    store.dispatch(action);
+    new Promise((resolve, reject) => {
+      const res = {
+        data: ['hello', 'javascript', 'world'],
+        status: 200
+      };
+      resolve(res.data);
+    })
+      .then(response => {
+        console.log('response', response);
+        const action = actionCreators.initListAction(response);
+        store.dispatch(action);
+      });
   }
 
   handleInputChange (e) {
