+++ src/store/index.js
@@ -1,28 +1,9 @@
-import { createStore, applyMiddleware, compose } from 'redux';
+import { createStore } from 'redux';
 import reducer from './reducer';
-import thunk from 'redux-thunk';
 
-// const store = createStore(
-//   reducer,
-//   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
-// );
-
-/**
- * redux-thunk中间件引入 与 redux引入 (github: redux-devtools-extension)
- * redux-thunk 使 actionCreator 函数的返回值可以是一个函数。
- */
-
-const composeEnhancers =
-  typeof window === 'object' &&
-  window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ ?   
-    window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({
-      // Specify extension’s options like name, actionsBlacklist, actionsCreators, serialize...
-    }) : compose;
-
-const enhancer = composeEnhancers(
-  applyMiddleware(thunk),
-  // other store enhancers if any
+const store = createStore(
+  reducer,
+  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
 );
-const store = createStore(reducer, enhancer);
 
 export default store;
